Add activity feed API (#4)
### TL;DR

This PR implements a new feature in the application.

### What changed?

A new feature was added to the application to improve user experience.

### How to test?

To test this feature, navigate to the feature section in the application and try using it.

### Why make this change?

This change was made to improve the overall user experience and functionality of the application.

---